+++ packages/article-byline/article-byline.stories.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Text } from "react-native";
 import { storiesOf } from "@storybook/react-native";
 import { checkA11y } from "@storybook/addon-a11y";
-import { fonts } from "@times-components/styleguide";
+import { fonts, colours } from "@times-components/styleguide";
 import ArticleByline from "./article-byline";
 
 const authorsAST = require("./fixtures/authors.json");
@@ -10,14 +10,14 @@ const authorsAST = require("./fixtures/authors.json");
 const bylineStyles = {
   fontSize: 13,
   fontFamily: fonts.supporting,
-  color: "#696969",
+  color: colours.functional.doveGrey,
   lineHeight: 13,
   flexDirection: "row"
 };
 
 const bylineLinkStyles = {
   link: {
-    color: "red",
+    color: colours.functional.azureBlue,
     textDecorationLine: "none"
   }
 };