+++ packages/article/styles/article-body/responsive.web.js
@@ -1,5 +1,6 @@
 import { View } from "react-native";
 import withResponsiveStyles from "@times-components/responsive-styles";
+import { colours } from "@times-components/styleguide";
 import config from "../responsive-config";
 
 /* --- Body --- */
@@ -12,7 +13,7 @@ export const ParagraphContainer = withResponsiveStyles("div", {
 
 export const Paragraph = withResponsiveStyles("p", {
   base: () => `
-    color: "#333333";
+    color: ${colours.functional.mineShaftGrey};
     font-family: "TimesDigitalW04-Regular";
     line-height: 26px;
     font-size: 17px;
@@ -28,7 +29,7 @@ export const Paragraph = withResponsiveStyles("p", {
 
 export const LinkStyle = withResponsiveStyles("a", {
   base: () => `
-    color: #069;
+    color: ${colours.functional.azureBlue};
     font-family: "TimesDigitalW04-Regular";
     line-height: 26px;
     font-size: 17px;