+++ packages/card/__tests__/web/card.web.test.js
@@ -1,7 +1,5 @@
 import React from "react";
-import { shallow } from "enzyme";
-import renderer from "react-test-renderer";
-import "jest-styled-components";
+import { shallow, mount } from "enzyme";
 import Card from "../../card";
 
 const cardProps = {
@@ -11,75 +9,70 @@ const cardProps = {
   },
   imageRatio: 1.5,
   imageSize: 360,
-  showImage: true
+  showImage: true,
+  tabletChildRatio: 2.7
 };
 
-describe("Card test on web", () => {
-  it("renders", () => {
-    const tree = renderer
-      .create(
-        <Card {...cardProps}>
-          <span>A card</span>
-        </Card>
-      )
-      .toJSON();
+describe("Card tests on web", () => {
+  it("should render a card", () => {
+    const component = shallow(
+      <Card {...cardProps}>
+        <span>A card</span>
+      </Card>
+    );
 
-    expect(tree).toMatchSnapshot();
+    expect(component).toMatchSnapshot("1. Render a Card");
   });
 
-  it("renders loading state", () => {
-    const tree = renderer
-      .create(
-        <Card {...cardProps} isLoading>
-          <span>A loading state</span>
-        </Card>
-      )
-      .toJSON();
+  it("should render the loading state", () => {
+    const component = mount(
+      <Card {...cardProps} isLoading>
+        <span>Loading state</span>
+      </Card>
+    );
 
-    expect(tree).toMatchSnapshot();
+    expect(component).toMatchSnapshot("2. Render the loading state");
   });
 
-  it("renders without image", () => {
+  it("should render without an image", () => {
     const noImageProps = Object.assign({}, cardProps, {
       image: null
     });
-    const tree = renderer
-      .create(
-        <Card {...noImageProps}>
-          <span>No image</span>
-        </Card>
-      )
-      .toJSON();
+    const component = shallow(
+      <Card {...noImageProps}>
+        <span>No image</span>
+      </Card>
+    );
 
-    expect(tree).toMatchSnapshot();
+    expect(component).toMatchSnapshot("3. Renders without an image");
   });
 
-  it("renders without image url", () => {
+  it("should render without image url", () => {
     const noImageProps = Object.assign({}, cardProps, {
       image: {
         uri: null
       }
     });
-    const tree = renderer
-      .create(
-        <Card {...noImageProps}>
-          <span>No URL</span>
-        </Card>
-      )
-      .toJSON();
+    const component = shallow(
+      <Card {...noImageProps}>
+        <span>No URL</span>
+      </Card>
+    );
 
-    expect(tree).toMatchSnapshot();
+    expect(component).toMatchSnapshot("4. Renders without an image URL");
   });
 
-  it("re-renders when image uri changes", () => {
+  it("should re-render when image uri changes", () => {
     const component = shallow(
       <Card {...cardProps}>
         <span>Some text</span>
       </Card>
     );
 
+    const testUri = "http://foo";
+
     component.setProps({
-      image: { uri: "http://foo" }
+      image: { uri: testUri }
     });
 
     expect(
@@ -87,18 +80,20 @@ describe("Card test on web", () => {
         .find("TimesImage")
         .at(0)
         .props().uri
-    ).toMatchSnapshot();
+    ).toEqual(`${testUri}&resize=${cardProps.imageSize}`);
   });
 
-  it("re-renders when image size changes", () => {
+  it("should re-render when image size changes", () => {
     const component = shallow(
       <Card {...cardProps}>
         <span>Some content</span>
       </Card>
     );
 
+    const testImageSize = 620;
+
     component.setProps({
-      imageSize: 620
+      imageSize: testImageSize
     });
 
     expect(
@@ -106,10 +101,10 @@ describe("Card test on web", () => {
         .find("TimesImage")
         .at(0)
         .props().uri
-    ).toMatchSnapshot();
+    ).toEqual(`${cardProps.image.uri}&resize=${testImageSize}`);
   });
 
-  it("re-renders when is loading changes", () => {
+  it("should re-render after showing the loading state", () => {
     const component = shallow(
       <Card {...cardProps} isLoading>
         <span>Re-render me</span>
@@ -120,10 +115,12 @@ describe("Card test on web", () => {
       isLoading: false
     });
 
-    expect(component.find("TimesImage").at(0)).toMatchSnapshot();
+    expect(component).toMatchSnapshot(
+      "5. Re-renders after showing the loading state"
+    );
   });
 
-  it("does not re-render fixed prop changes", () => {
+  it("should not re-render when image ratio changes", () => {
     const component = shallow(
       <Card {...cardProps}>
         <span>Do not re-render me</span>
@@ -134,6 +131,8 @@ describe("Card test on web", () => {
       imageRatio: 2
     });
 
-    expect(component.find("TimesImage").at(0)).toMatchSnapshot();
+    expect(component).toMatchSnapshot(
+      "6. Renders Card normally and does not re-render"
+    );
   });
 });