+++ packages/article/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.42.1"></a>
+## [0.42.1](https://github.com/newsuk/times-components/compare/@times-components/article@0.42.0...@times-components/article@0.42.1) (2018-03-02)
+
+
+
+
+**Note:** Version bump only for package @times-components/article
+
 <a name="0.42.0"></a>
 # [0.42.0](https://github.com/newsuk/times-components/compare/@times-components/article@0.40.5...@times-components/article@0.42.0) (2018-03-01)